requirements.txt
fastapi==0.115.6
passlib==1.7.4
pydantic==2.10.4
pydantic[email]
python-jose==3.3.0
sqlalchemy==2.0.36
uvicorn==0.34.0
